Make Handlers out of CreateGame and LoadGame.
